requirements.txt
tqdm
translate
pyyaml
ffmpeg-python
